requirements.txt
streamlit
langchain-google-genai
google-genai
python-dotenv
pypdf
python-docx